# docs/architecture/06-api-contracts/openapi-spec.yaml
openapi: 3.0.3
info:
  title: KMS API
  description: |
    Knowledge Management System API for file indexing, search, and duplicate detection.

    ## Authentication

    All endpoints require authentication via:
    - **API Key**: `X-API-Key` header
    - **JWT Token**: `Authorization: Bearer <token>` header

    ## Rate Limiting

    Default rate limits:
    - Free tier: 60 requests/minute
    - Pro tier: 300 requests/minute
    - Enterprise: 1000 requests/minute

    Rate limit headers are included in all responses.
  version: 1.0.0
  contact:
    name: KMS API Support
    email: api-support@kms.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development
  - url: https://staging-api.kms.example.com/api/v1
    description: Staging
  - url: https://api.kms.example.com/api/v1
    description: Production
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  expires_in:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /sources:
    get:
      tags: [Sources]
      summary: List sources
      parameters:
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/SourceType'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/SourceStatus'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Sources list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourcesListResponse'

    post:
      tags: [Sources]
      summary: Connect source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSourceRequest'
      responses:
        '201':
          description: Source connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SourceType:
      type: string
      enum: [google_drive, local_fs, onedrive, dropbox]
    SourceStatus:
      type: string
      enum: [active, syncing, error, disconnected, paused]
    Source:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/SourceType'
        status:
          $ref: '#/components/schemas/SourceStatus'
        root_path:
          type: string
        config:
          type: object
        stats:
          $ref: '#/components/schemas/SourceStats'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SourceStats:
      type: object
      properties:
        files_indexed:
          type: integer
        total_size_bytes:
          type: integer
        last_scan_at:
          type: string
          format: date-time
        next_scan_at:
          type: string
          format: date-time
    SourcesListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Source'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateSourceRequest:
      type: object
      required: [name, type, config]
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/SourceType'
        config:
          type: object
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        cursor:
          type: string
        next_cursor:
          type: string
        has_more:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
            request_id:
              type: string
            timestamp:
              type: string
              format: date-time